Drop inclusion of tools/prewin.h and postwin.h from sunjavaplugin.cxx
That file compiles fine without them. These headers are in tools,
which hasn't been built and delivered yet when sunjvmfwk is built. We
can't add tools to the dependencies of jvmfwk as that would introduce
circularity.

As such, I see no reason why prewin.h and postwin.h are in tools. I
think they should be in solenv, like premac.h and postmac.h. It would
perhaps be an Easy Hack to move them, and drop the "tools/" from all
lines including them?