Rehash cross-compilation ideas
Drop the TARGETPLATFORM=BUILD indication of stuff that is to be built
for the build platform. I will handle the split of stuff built for the
build or host platforms differently.

Add explicit rules to do nothing for the cross-compilation case, but
likely even that will be unnecessary in the case of complete modules
like xml2cmp. I will just mark modules that are for the build platform
only with an own flag in BUILD_TYPE.