Cross-compilation fix
checknewapi.pl can run regview fine on Windows even if the name
doesn't include the .exe suffix. So drop the use of EXECPOST and then
this will work also in the case of cross-compilation to Windows, where
EXECPOST is .exe for HOST, but we aren't running a HOST regview but a
BUILD one.