Just make -lstore work for MinGW as it does for other Unix-style compilers
I don't really see the point in having the UDK_MAJOR in the import
library name. We don't do that for any Unix either, just -lstore is
used. It isn't as if it would make any sense in keeping a mix of
(import) libraries for incompatible versions/builds of the LO (SDK) in
the same location anyway. The simpler, the better.