iOS hacking for cppunittester
We don't need sal/main.h in salinit.cxx as far as I can see.

Add stuff to sal/main.h so that we can link an iOS executable that
actually will run, and hack the makefile so that cppunittester builds
successfully.

Of course, a cppunittester executable built here won't be usable for
anything interesting as we don't build any dynamic libraries of
LibreOffice code for iOS, so there is nothing to pass on the command
line to load. But hey, baby steps, experimentation, doing work in the
public, commit often, etc...